requirements.txt
streamlit==1.31.0
anthropic==0.18.1
langchain==0.1.9
langchain-community==0.0.24
langchain-text-splitters==0.0.1
chromadb==0.4.22
sentence-transformers==2.3.1
python-dotenv==1.0.1